Move rlp::encode(vector) into encode_vector.hpp (#492)
* Move rlp::encode(vector) into encode_vector.hpp

* Delete RLPVectorCoding.hpp

* Simplify duplicate ommer check per Greg's suggestion

* Reserve capacity to avoid possible reallocation (kudos to Greg)

* More precise reserve in encode(vector)

* Fix for Wasm where size_t is uint32_t, not uint64_t

* Bump resource_class for linux-clang-coverage
