+++ core/silkworm/consensus/base/engine.cpp
@@ -16,10 +16,11 @@
 
 #include "engine.hpp"
 
-#include <silkworm/common/as_range.hpp>
 #include <silkworm/chain/difficulty.hpp>
 #include <silkworm/chain/protocol_param.hpp>
+#include <silkworm/common/as_range.hpp>
 #include <silkworm/crypto/ecdsa.hpp>
+#include <silkworm/rlp/encode_vector.hpp>
 #include <silkworm/trie/vector_root.hpp>
 
 namespace silkworm::consensus {
@@ -68,7 +69,7 @@ ValidationResult ConsensusEngineBase::pre_validate_block(const silkworm::Block& 
             return ValidationResult::kNotAnOmmer;
         }
 
-        if (as_range::any_of(old_ommers, [&ommer](const BlockHeader& old_ommer) -> bool { return old_ommer == ommer; })) {
+        if (as_range::find(old_ommers, ommer) != old_ommers.end()) {
             return ValidationResult::kDuplicateOmmer;
         }
     }
