+++ core/silkworm/types/log.cpp
@@ -17,7 +17,7 @@
 #include "log.hpp"
 
 #include <silkworm/common/util.hpp>
-#include <silkworm/rlp/encode.hpp>
+#include <silkworm/rlp/encode_vector.hpp>
 
 namespace silkworm ::rlp {
