+++ core/silkworm/types/receipt.cpp
@@ -17,7 +17,7 @@
 #include "receipt.hpp"
 
 #include <silkworm/common/util.hpp>
-#include <silkworm/rlp/encode.hpp>
+#include <silkworm/rlp/encode_vector.hpp>
 
 namespace silkworm::rlp {
